refactor: change model product price column type to Decimal

+++ back-end/src/database/models/products.js
@@ -12,7 +12,7 @@ const makeProducts = (sequelize, DataTypes) => {
     },
     price: {
       allowNull: false,
-      type: DataTypes.FLOAT
+      type: DataTypes.DECIMAL(10, 2)
     },
     urlImage: {
       allowNull: false,
